refactor to use custom hook and add click off handler
