+++ src/components/Message/hooks/index.js
@@ -4,6 +4,7 @@ export * from './useDeleteHandler';
 export * from './useEditHandler';
 export * from './useFlagHandler';
 export * from './useMentionsHandler';
+export * from './useMobilePress';
 export * from './useMuteHandler';
 export * from './useOpenThreadHandler';
 export * from './usePinHandler';
